backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography>=41.0.0

# Data Validation
pydantic==2.5.0
pydantic[email]==2.5.0
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment & Configuration
python-dotenv==1.0.0

# File Upload & Form Data Support
python-multipart==0.0.6

# Gmail API Integration
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
